requirements-test.txt
# Test requirements for PetTracer Home Assistant integration
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
pytest-cov>=4.1.0
